refactor(web): extract thread timeline derivation into pure module
Move the pure derive steps (display message attachment mapping, preview
handoff/optimistic merge, turn diff summary index, revert turn counts)
out of ChatView's inline useMemos into threadTimelineModel.ts so the
thread-state -> timeline pipeline can run outside the app shell (perf
harnesses, tests). Behavior-identical.
